requirements.txt
boto3
pandas
PyPDF2
python-pptx
python-docx
qdrant-client
langchain
langchain-community
opensearch-py
fastapi
uvicorn
pydantic
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
openai
langchain-openai
